requirements.txt
flask==2.0.1
werkzeug==2.0.1
requests==2.26.0
python-dotenv==0.19.0
boto3==1.18.44
gunicorn==20.1.0
SQLAlchemy==1.4.23
PyMySQL==1.0.2
cryptography==41.0.7
pyOpenSSL==23.2.0
urllib3==1.26.15
botocore==1.21.44 